tic_tac_toe/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2==2.9.1